Reorganized routes and added pages
